+++ backend/app/schemas/tax.py
@@ -1,19 +1,25 @@
 from datetime import date
+from decimal import Decimal
 
 from pydantic import BaseModel
+
+
+# H-ARCH-2: Decimal statt float bei Geldbetraegen. Pydantic v2 serialisiert
+# Decimal standardmaessig als JSON-String — Frontend formatAmount() macht
+# parseFloat, daher kompatibel ohne Anpassung.
 
 
 class TaxCategorySummary(BaseModel):
     category: str
     label: str
     document_count: int
-    total_amount: float
+    total_amount: Decimal
 
 
 class TaxYearSummary(BaseModel):
     year: int
     total_documents: int
-    total_amount: float
+    total_amount: Decimal
     categories: list[TaxCategorySummary]
     warnings: list[str] = []
 
@@ -37,7 +43,7 @@ class TaxDocumentItem(BaseModel):
     title: str
     document_type: str
     document_date: date | None = None
-    amount: float | None = None
+    amount: Decimal | None = None
     currency: str = "EUR"
     issuer: str | None = None
     tax_category: str | None = None
